refactor: groups and renames detail files
